02
Here, one can input any random texts and check whether the text is Palindrome or not. A palindrome is a word, sentence, verse, or even number that reads the same backward or forward.

# condition
This program is written in such a beautiful manner that while checking the Palindrome, the spaces and special characters are ignored.